+++ samples/order_service/src/order_service/models/entities/__init__.py
@@ -1,0 +1,19 @@
+# Copyright 2026 Firefly Software Foundation.
+# Licensed under the Apache License, Version 2.0.
+from order_service.models.entities.v1.order_entity import (
+    Order,
+    OrderCancelled,
+    OrderInventoryReserved,
+    OrderPaid,
+    OrderPlaced,
+    OrderShipped,
+)
+
+__all__ = [
+    "Order",
+    "OrderCancelled",
+    "OrderInventoryReserved",
+    "OrderPaid",
+    "OrderPlaced",
+    "OrderShipped",
+]